Moving the runner out of __init__
